Added support for new version of Dlink DWA-171 (version C)
There seems to be a new version of the Dlink DWA-171 WiFi dongle
available which does not have the same USB ID (2001:331d) and uses
the C version of the driver as the old ones (2001:3314). Added this
new USB product ID to the list and the new dongles are working.